Move resize handler into lib/libraries/default.js
