+++ lib/index.js
@@ -5,7 +5,7 @@
 
   // Visualization Libraries
   var libraries = {
-    'default': require('./libraries/default')()
+    'default': require('./libraries/default')(Dataviz)
     // TODO: add remaining
   };
 
@@ -13,8 +13,6 @@
   var each = require('./utils/each'),
       extend = require('./utils/extend'),
       isDateString = require('./utils/assert-date-string');
-
-  handleWindowResize();
 
   // Constructor
   function Dataviz(){
@@ -455,34 +453,6 @@
     else {
       fn();
     }
-  }
-
-  function bindResizeListener(fn){
-    if ('undefined' === typeof window) return;
-    window.onresize = window.resize = function(){};
-    if (window.addEventListener) {
-      window.addEventListener('resize', fn, true);
-    }
-    else if (window.attachEvent) {
-      window.attachEvent('onresize', fn);
-    }
-  }
-
-  function handleWindowResize(){
-    var timer, delay;
-    bindResizeListener(function(){
-      if (timer) {
-        clearTimeout(timer);
-      }
-      delay = (Dataviz.visuals.length > 12) ? 1000 : 250;
-      timer = setTimeout(function(){
-        each(Dataviz.visuals, function(chart){
-          if (chart.view._artifacts.c3) {
-              chart.view._artifacts.c3.resize();
-          }
-        });
-      }, delay);
-    });
   }
 
   // Exports
