+++ lib/libraries/default.js
@@ -7,7 +7,22 @@ var each = require('../utils/each'),
 
 var types = {};
 
-function initialize(){
+function initialize(lib){
+  var timer, delay;
+  bindResizeListener(function(){
+    if (timer) {
+      clearTimeout(timer);
+    }
+    delay = (lib.visuals.length > 12) ? 1000 : 250;
+    timer = setTimeout(function(){
+      each(lib.visuals, function(chart){
+        if (chart.view._artifacts.c3) {
+            chart.view._artifacts.c3.resize();
+        }
+      });
+    }, delay);
+  });
+
   defineC3();
   defineMessage();
   defineMetric();
@@ -130,6 +145,17 @@ function defineC3(){
       }
     };
   });
+}
+
+function bindResizeListener(fn){
+  if ('undefined' === typeof window) return;
+  window.onresize = window.resize = function(){};
+  if (window.addEventListener) {
+    window.addEventListener('resize', fn, true);
+  }
+  else if (window.attachEvent) {
+    window.attachEvent('onresize', fn);
+  }
 }
 
 function getDateFormatDefault(a, b){
